Make SolarRealPowerLimit a float64
The PW can expose a value equivalent to math.MaxFloat64
(1.7976931348623157e+308), which can't be JSON-unmarshalled into a
float32.

Corrects errors of the form:

Sep 01 22:22:22 hostname powerwall_exporter[1234]: time="2023-09-03T22:36:07Z" level=error msg="Error fetching system_status info" err="json: cannot unmarshal number 1.7976931348623157e+308 into Go struct field SystemStatusData.solar_real_power_limit of type float32"

+++ system.go
@@ -66,7 +66,7 @@ type SystemStatusData struct {
 	SmartInvDeltaP             float32         `json:"smart_inv_delta_p"`
 	SmartInvDeltaQ             float32         `json:"smart_inv_delta_q"`
 	LastToggleTimestamp        time.Time       `json:"last_toggle_timestamp"`
-	SolarRealPowerLimit        float32         `json:"solar_real_power_limit"`
+	SolarRealPowerLimit        float64         `json:"solar_real_power_limit"`
 	Score                      float32         `json:"score"`
 	BlocksControlled           int             `json:"blocks_controlled"`
 	Primary                    bool            `json:"primary"`
